refactor: extract supported-languages parsing into a shared helper
The Steam supported_languages parsing was duplicated verbatim between the
big picture SupportedLanguages component and the desktop
GameLanguageSection. Extract it into a shared parseSupportedLanguages
helper used by both, removing the duplication.

The input is Steam's supported_languages HTML string: a comma separated
list terminated by a <br>, where a trailing <strong>*</strong> marks the
languages that ship full audio. The helper takes the segment before the
<br>, strips that marker off each entry and reports it as hasAudio.
