+++ src/shared/index.ts
@@ -31,6 +31,7 @@ export * from "./artwork-resolver";
 export * from "./download-directories";
 export * from "./html-sanitizer";
 export * from "./language-flags";
+export * from "./supported-languages";
 export * from "./use-hls-video";
 export * from "./retroarch-platform";
 export * from "./tracker-list";
